change color and font size

+++ klib/string.c
@@ -1,7 +1,7 @@
 #include <klib.h>
 
-size_t strlen(const char *s) {
-  size_t sz = 0;
+int strlen(const char *s) {
+  int sz = 0;
   for (; *s; s ++) {
     sz ++;
   }
